The game expected the constructer to use doubles not floats ig

+++ src/main/java/brokendoop/doopmod/entity/EntityLaserdustFX.java
@@ -7,19 +7,19 @@ import net.minecraft.core.world.World;
 public class EntityLaserdustFX extends EntityFX {
 	float field_673_a;
 
-	public EntityLaserdustFX(World world, double d, double d1, double d2, float f, float f1, float f2) {
+	public EntityLaserdustFX(World world, double d, double d1, double d2, double f, double f1, double f2) {
 		this(world, d, d1, d2, 1.0F, f, f1, f2);
 	}
 
-	public EntityLaserdustFX(World world, double d, double d1, double d2, float f, float red, float green, float blue) {
+	public EntityLaserdustFX(World world, double d, double d1, double d2, double f, double red, double green, double blue) {
 		super(world, d, d1, d2, 0.0, 0.0, 0.0);
 		this.xd *= 0.10000000149011612;
 		this.yd *= 0.10000000149011612;
 		this.zd *= 0.10000000149011612;
 		float f4 = (float)Math.random() * 0.4F + 0.6F;
-		this.particleRed = ((float)(Math.random() * 0.20000000298023224) + 0.8F) * red * f4;
-		this.particleGreen = ((float)(Math.random() * 0.20000000298023224) + 0.8F) * green * f4;
-		this.particleBlue = ((float)(Math.random() * 0.20000000298023224) + 0.8F) * blue * f4;
+		this.particleRed = (float) (((Math.random() * 0.20000000298023224) + 0.8F) * red * f4);
+		this.particleGreen = (float) (((Math.random() * 0.20000000298023224) + 0.8F) * green * f4);
+		this.particleBlue = (float) (((Math.random() * 0.20000000298023224) + 0.8F) * blue * f4);
 		this.particleScale *= 0.75F;
 		this.particleScale *= f;
 		this.field_673_a = this.particleScale;
